% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmi.cad}
\alias{bmi.cad}
\title{Effect of Body Mass Index (BMI) on Coronary Artery Disease (CAD)}
\format{A \code{data.frame} with 1119 rows and 42 variables.}
\source{
\url{https://github.com/qingyuanzhao/mr.raps}
}
\usage{
data(bmi.cad)
}
\description{
It contains independent datasets from three genome-wide association studies (GWASs):
\enumerate{
\item Exposure dataset: A GWAS for BMI in round 2 of the UK BioBank (sample size: 336,107), \url{http://www.nealelab.is/uk-biobank}.
\item Outcome dataset: A GWAS A GWAS for CAD from the CARDIoGRAMplusC4D consortium  (sample size: ~185,000), with genotype imputation using the 1000 Genome Project, (PubMed 26343387).
\item Selection dataset: A GWAS for BMI in the Japanese population (sample size: 173,430), (PubMed 28892062).
}
}
\keyword{datasets}
