% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_gen_individual}
\alias{data_gen_individual}
\title{Summary Statistics from Simulated Individual-level data}
\usage{
data_gen_individual(case = c("case4", "case5", "case6", "case7"),
  true_var = FALSE)
}
\arguments{
\item{case}{Simulation scenario used in Section 5.3 in Ye et al., (2020).}

\item{true_var}{Do se.exposure, se.outcome, se.selection equal the true standard deviations or the estimated standard errors. Default is FALSE.}
}
\value{
A data frame
}
\description{
Summary Statistics from Simulated Individual-level data
}
\references{
Ting Ye, Jun Shao, Hyunseung Kang (2020). Debiased Inverse-Variance Weighted Estimator in Two-Sample Summary-Data Mendelian Randomization.\url{https://arxiv.org/abs/1911.09802}.
}
