% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_gen_summary}
\alias{data_gen_summary}
\title{Summary Statistics Simulated from the BMI-CAD Dataset}
\usage{
data_gen_summary(case = c("case1", "case2", "case3", "case3_pleiotropy"))
}
\arguments{
\item{case}{Simulation scenario used in Section 5.1 in Ye et al., (2020).}
}
\value{
A data frame
}
\description{
Summary Statistics Simulated from the BMI-CAD Dataset
}
\references{
Ting Ye, Jun Shao, Hyunseung Kang (2020). Debiased Inverse-Variance Weighted Estimator in Two-Sample Summary-Data Mendelian Randomization.\url{https://arxiv.org/abs/1911.09802}.
}
