% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mr.divw}
\alias{mr.divw}
\title{Main function for dIVW}
\usage{
mr.divw(
  beta.exposure,
  beta.outcome,
  se.exposure,
  se.outcome,
  alpha = 0.05,
  pval.selection = NULL,
  lambda = 0,
  over.dispersion = FALSE,
  diagnostics = FALSE,
  overlap = FALSE,
  gen_cor = 0
)
}
\arguments{
\item{beta.exposure}{A vector of SNP effects on the exposure vairable, usually obtained from a GWAS}

\item{beta.outcome}{A vector of SNP effects on the outcome vairable, usually obtained from a GWAS}

\item{se.exposure}{A vecor of standard errors of \code{beta.exposure}}

\item{se.outcome}{A vector of standard errors of \code{beta.outcome}}

\item{alpha}{Confidence interval has level 1-alpha}

\item{pval.selection}{A vector of p-values calculated based on the selection dataset that is used for IV selection. It is not required when lambda=0}

\item{lambda}{The specified z-score threhold. Default is 0 (without thresholding)}

\item{over.dispersion}{Should the model consider balanced horizontal pleiotropy. Default is FALSE}

\item{diagnostics}{Should the function returns the q-q plot for assumption diagnosis. Default is FALSE}

\item{overlap}{Should the model consider overlapping exposure and outcome datasets. Default is FALSE}

\item{gen_cor}{If overlap = TRUE, provide an estimate of the correlation between the effect of the genetic variants on the exposure and the outcome. Default value is 0, meaning that the exposure and outcome datasets are non-overlapping.}
}
\value{
A list
\describe{
\item{beta.hat}{Estimated causal effect}
\item{beta.se}{Standard error of \code{beta.hat}}
\item{condition}{A measure that needs to be large for reliable asymptotic approximation based on the dIVW estimator. It is recommended to be greater than 20}
\item{tau.square}{Overdispersion parameter if \code{over.dispersion=TRUE}}
\item{n.IV}{Number of IVs used in the dIVW estimator}
\item{IV}{IVs that are used in the dIVW estimator}
}
}
\description{
Main function for dIVW
}
\examples{

data(bmi.cad)
with(bmi.cad,
  mr.divw(beta.exposure, beta.outcome, se.exposure, se.outcome, diagnostics=TRUE)
)

}
\references{
Ting Ye, Jun Shao, Hyunseung Kang (2020). Debiased Inverse-Variance Weighted Estimator in Two-Sample Summary-Data Mendelian Randomization.\url{https://arxiv.org/abs/1911.09802}.
}
