% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mr.eo}
\alias{mr.eo}
\title{MR-EO Algorithm to Adaptively Find the Optimal Z-score Threhold.}
\usage{
mr.eo(
  lambda.start,
  beta.exposure,
  beta.outcome,
  se.exposure,
  se.outcome,
  pval.selection,
  over.dispersion = FALSE,
  max_opt_iter = 5
)
}
\arguments{
\item{lambda.start}{Initial value for lambda (the z-score threshold).}

\item{beta.exposure}{A vector of SNP effects on the exposure vairable, usually obtained from a GWAS}

\item{beta.outcome}{A vector of SNP effects on the outcome vairable, usually obtained from a GWAS}

\item{se.exposure}{A vecor of standard errors of \code{beta.exposure}}

\item{se.outcome}{A vector of standard errors of \code{beta.outcome}}

\item{pval.selection}{A vector of p-values calculated based on the selection dataset that is used for IV selection}

\item{over.dispersion}{Should the model consider balanced horizontal pleiotropy. Default is FALSE}

\item{max_opt_iter}{Maximum number of iterations. Default is 5}
}
\value{
A list
\describe{
\item{lambda.opt}{Optimal z-socre threshold}
\item{n.iter}{Number of iterations to find \code{lambda.opt}}
}
}
\description{
MR-EO Algorithm to Adaptively Find the Optimal Z-score Threhold.
}
\details{
\code{mr.eo} is an adaptive algorithm that finds the optimal z-socre threshold that leads to the dIVW estimator with the smallest variance.
}
\examples{

df <- data_gen_summary("case1")
lambda.opt <- with(df,
  mr.eo(0, beta.exposure, beta.outcome, se.exposure, se.outcome,
    pval.selection)$lambda.opt
)
with(df,
  mr.divw(beta.exposure, beta.outcome, se.exposure, se.outcome,
    pval.selection = pval.selection, lambda = lambda.opt)
)

data(bmi.cad)
lambda.opt <- with(bmi.cad,
  mr.eo(0, beta.exposure, beta.outcome, se.exposure, se.outcome,
    pval.selection)$lambda.opt
)
with(bmi.cad,
  mr.divw(beta.exposure, beta.outcome, se.exposure, se.outcome,
    pval.selection = pval.selection, lambda = lambda.opt)
)

}
\references{
Ting Ye, Jun Shao, Hyunseung Kang (2020). Debiased Inverse-Variance Weighted Estimator in Two-Sample Summary-Data Mendelian Randomization.\url{https://arxiv.org/abs/1911.09802}.
}
