% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmr.divw.overlap.R
\name{mvmr.divw.overlap}
\alias{mvmr.divw.overlap}
\title{Perform debiased adjusted inverse-variance weighted (adIVW) estimator for summary-data multivariable Mendelian randomization allowing for overlapping exposure and outcome datasets}
\usage{
mvmr.divw.overlap(
  beta.exposure,
  se.exposure,
  beta.outcome,
  se.outcome,
  gen_cor = NULL,
  phi_cand = 0
)
}
\arguments{
\item{beta.exposure}{A data.frame or matrix. Each row contains the estimated marginal effect of a SNP on K exposures, usually obtained from a GWAS}

\item{se.exposure}{A data.frame or matrix of estimated standard errors of beta.exposure}

\item{beta.outcome}{A vector of the estimated marginal effect of a SNP on outcome, usually obtained from a GWAS}

\item{se.outcome}{A vector of estimated standard errors of beta.outcome}

\item{gen_cor}{Provide a (K+1)-by-(K+1) matrix for the estimated shared correlation matrix between the effect of the genetic variants on each exposure, where K is the number of exposure and the last index position corresponds to the outcome. The correlations can either be estimated, be assumed to be zero, or fixed at zero using non-overlapping samples of each exposure GWAS. Default input is NULL, meaning that an identity matrix is used as the correlation matrix.}

\item{phi_cand}{A vector of tuning parameters for adIVW estimator. Default is 0, meaning that dIVW estimator is performed. To use the recommended set for the tuning parameter, simply set phi_cand = NULL.}
}
\value{
A list with elements
\item{beta.hat}{Estimated direct effects of each exposure on the outcome}
\item{beta.se}{Estimated standard errors of beta.hat}
\item{iv_strength_parameter}{The minimum eigenvalue of the sample IV strength matrix, which quantifies the IV strength in the sample}
\item{phi_selected}{The selected tuning parameter for the adIVW estimator}
}
\description{
Perform debiased adjusted inverse-variance weighted (adIVW) estimator for summary-data multivariable Mendelian randomization allowing for overlapping exposure and outcome datasets
}
\examples{
library(MVMR)
data("rawdat_mvmr")
beta.exposure <- rawdat_mvmr[,c("LDL_beta","HDL_beta","Trg_beta")]
se.exposure <- rawdat_mvmr[,c("LDL_se","HDL_se","Trg_se")]
beta.outcome <- rawdat_mvmr$SBP_beta
se.outcome <- rawdat_mvmr$SBP_se
P <- matrix(0.3, nrow = 4, ncol = 4)
diag(P) <- 1
mvmr.divw.overlap(beta.exposure = beta.exposure,
se.exposure = se.exposure,
beta.outcome = beta.outcome,
se.outcome = se.outcome,
gen_cor = P)

}
