% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmr.divw.R
\name{mvmr.divw}
\alias{mvmr.divw}
\title{Perform debiased adjusted inverse-variance weighted (adIVW) estimator for summary-data multivariable Mendelian randomization}
\usage{
mvmr.divw(
  beta.exposure,
  se.exposure,
  beta.outcome,
  se.outcome,
  gen_cor = NULL,
  phi_cand = 0,
  over.dispersion = FALSE
)
}
\arguments{
\item{beta.exposure}{A data.frame or matrix. Each row contains the estimated marginal effect of a SNP on K exposures, usually obtained from a GWAS}

\item{se.exposure}{A data.frame or matrix of estimated standard errors of beta.exposure}

\item{beta.outcome}{A vector of the estimated marginal effect of a SNP on outcome, usually obtained from a GWAS}

\item{se.outcome}{A vector of estimated standard errors of beta.outcome}

\item{phi_cand}{A vector of tuning parameters for adIVW estimator. Default is 0, meaning that dIVW estimator is performed. To use the recommended set for the tuning parameter, simply set phi_cand = NULL.}

\item{over.dispersion}{Should the model consider balanced horizontal pleiotropy. Default is FALSE}

\item{gen_cov}{If the exposure and outcome datasets are non-overlapping, then provide a K-by-K matrix for the estimated shared correlation matrix between the effect of the genetic variants on each exposure, where K is the number of exposure. The correlations can either be estimated, be assumed to be zero, or fixed at zero using non-overlapping samples of each exposure GWAS. Otherwise, provide a (K+1)-by-(K+1) correlation matrix where the last row and column corresponds to the SNP-outcome associations. Default input is NULL, meaning that an identity matrix is used as the correlation matrix.}
}
\value{
A list with elements
\item{beta.hat}{Estimated direct effects of each exposure on the outcome}
\item{beta.se}{Estimated standard errors of beta.hat}
\item{iv_strength_parameter}{The minimum eigenvalue of the sample IV strength matrix, which quantifies the IV strength in the sample}
\item{phi_selected}{The selected tuning parameter for the adIVW estimator}
\item{tau.square}{Overdispersion parameter if \code{over.dispersion=TRUE}}
}
\description{
Perform debiased adjusted inverse-variance weighted (adIVW) estimator for summary-data multivariable Mendelian randomization
}
\examples{
library(MVMR)
data("rawdat_mvmr")
beta.exposure <- rawdat_mvmr[,c("LDL_beta","HDL_beta","Trg_beta")]
se.exposure <- rawdat_mvmr[,c("LDL_se","HDL_se","Trg_se")]
beta.outcome <- rawdat_mvmr$SBP_beta
se.outcome <- rawdat_mvmr$SBP_se
P <- matrix(0.3, nrow = 3, ncol = 3)
diag(P) <- 1
mvmr.divw(beta.exposure = beta.exposure,
se.exposure = se.exposure,
beta.outcome = beta.outcome,
se.outcome = se.outcome,
gen_cor = P,
phi_cand = NULL,
over.dispersion = FALSE)

}
