% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmr.divw.R
\name{mvmr.divw}
\alias{mvmr.divw}
\title{Perform debiased adjusted inverse-variance weighted (adIVW) estimator for two-sample summary-data multivariable Mendelian randomization}
\usage{
mvmr.divw(
  beta.exposure,
  se.exposure,
  beta.outcome,
  se.outcome,
  phi_cand = 0,
  over.dispersion = FALSE,
  overlap = FALSE,
  gen_cor = NULL
)
}
\arguments{
\item{beta.exposure}{A data.frame or matrix. Each row contains the estimated marginal effect of a SNP on K exposures, usually obtained from a GWAS}

\item{se.exposure}{A data.frame or matrix of estimated standard errors of beta.exposure}

\item{beta.outcome}{A vector of the estimated marginal effect of a SNP on outcome, usually obtained from a GWAS}

\item{se.outcome}{A vector of estimated standard errors of beta.outcome}

\item{phi_cand}{A vector of tuning parameters for adIVW estimator. Default is 0, meaning that dIVW estimator is performed. To use the recommended set for the tuning parameter, simply set phi_cand = NULL.}

\item{over.dispersion}{Should the model consider balanced horizontal pleiotropy? Default is FALSE.}

\item{overlap}{Should the model consider overlapping exposure and outcome datasets? Default is FALSE.}

\item{gen_cor}{If overlap = FALSE, provide a K-by-K matrix for the estimated shared correlation matrix between the effect of the genetic variants on each exposure, where K is the number of exposure. If overlap = TRUE, provide a (K+1)-by-(K+1) matrix for the estimated shared correlation matrix between the effect of the genetic variants on each exposure and the outcome, where the last index position corresponds to the outcome.  The correlations can either be estimated, be assumed to be zero, or fixed at zero. Default input is NULL, meaning that an identity matrix is used as the correlation matrix.}
}
\value{
A list with elements
\item{beta.hat}{Estimated direct effects of each exposure on the outcome}
\item{beta.se}{Estimated standard errors of beta.hat}
\item{iv_strength_parameter}{The minimum eigenvalue of the sample IV strength matrix, which quantifies the IV strength in the sample}
\item{phi_selected}{The selected tuning parameter for the adIVW estimator}
\item{tau.square}{Overdispersion parameter if \code{over.dispersion=TRUE}}
}
\description{
Perform debiased adjusted inverse-variance weighted (adIVW) estimator for two-sample summary-data multivariable Mendelian randomization
}
\examples{
data("hdl_subfractions")
# We are going to estimate the effect of S-HDL-P on the risk of CAS with adjustments of HDL, LDL, and TG levels
beta.exposure <- hdl_subfractions$data[,c("gamma_exp1","gamma_exp2","gamma_exp3","gamma_exp4")] # corresponding to SNP effects on respectively  HDL, LDL, TG, and S-HDL-P
se.exposure <- hdl_subfractions$data[,c("se_exp1","se_exp2","se_exp3","se_exp4")]
beta.outcome <- hdl_subfractions$data$gamma_out1
se.outcome <- hdl_subfractions$data$se_out1
P <- hdl_subfractions$cor.mat[c(1:4,10),c(1:4,10)] # make sure the last index corresponds to the outcome
mvmr.divw(beta.exposure = beta.exposure,
se.exposure = se.exposure,
beta.outcome = beta.outcome,
se.outcome = se.outcome,
gen_cor = P,
phi_cand = NULL,
over.dispersion = FALSE,
overlap = TRUE)

}
