% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmr.ivw.R
\name{mvmr.ivw}
\alias{mvmr.ivw}
\title{Perform inverse-variance weighted (IVW) estimator for two-sample summary-data multivariable Mendelian randomization}
\usage{
mvmr.ivw(beta.exposure, se.exposure, beta.outcome, se.outcome, gen_cor = NULL)
}
\arguments{
\item{beta.exposure}{A data.frame or matrix. Each row contains the estimated marginal effect of a SNP on K exposures, usually obtained from a GWAS}

\item{se.exposure}{A data.frame or matrix of estimated standard errors of beta.exposure}

\item{beta.outcome}{A vector of the estimated marginal effect of a SNP on outcome, usually obtained from a GWAS}

\item{se.outcome}{A vector of estimated standard errors of beta.outcome}

\item{gen_cor}{A K-by-K matrix for the estimated shared correlation matrix between the effect of the genetic variants on each exposure, where K is the number of exposure. The correlations can either be estimated, be assumed to be zero, or fixed at zero using non-overlapping samples of each exposure GWAS. Default input is NULL, meaning that an identity matrix is used as the correlation matrix.}
}
\value{
A list with elements
\item{beta.hat}{Estimated direct effects of each exposure on the outcome}
\item{beta.se}{Estimated standard errors of beta.hat}
\item{iv_strength_parameter}{The minimum eigenvalue of the sample IV strength matrix, which quantifies the IV strength in the sample}
}
\description{
Perform inverse-variance weighted (IVW) estimator for two-sample summary-data multivariable Mendelian randomization
}
\examples{
library(MVMR)
data("rawdat_mvmr")
beta.exposure <- rawdat_mvmr[,c("LDL_beta","HDL_beta","Trg_beta")]
se.exposure <- rawdat_mvmr[,c("LDL_se","HDL_se","Trg_se")]
beta.outcome <- rawdat_mvmr$SBP_beta
se.outcome <- rawdat_mvmr$SBP_se
P <- matrix(0.3, nrow = 3, ncol = 3)
diag(P) <- 1
mvmr.ivw(beta.exposure = beta.exposure,
se.exposure = se.exposure,
beta.outcome = beta.outcome,
se.outcome = se.outcome,
gen_cor = P)

}
